requirements.txt
argparse
jsons
PyYAML
glob2
regex
openai
